% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.r
\name{calibrate}
\alias{calibrate}
\title{Calibrate 14C age determinations}
\usage{
calibrate(age, error, curve = "SHcal13.txt", eps = 1e-05)
}
\arguments{
\item{age}{a numeric vector containing the mean 14C age for the samples.}

\item{error}{a numeric vector containing the error for each 14C
determination.}

\item{curve}{a string, currently only \code{"SHcal13.txt"} implemented.}

\item{eps}{the smallest probability to be considered as something. Values
under this number will not be included in the \code{pmf} output.}
}
\value{
a list of \code{pmf} objects.
}
\description{
Calibrate 14C age determinations
}
