% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_coreData.r
\docType{class}
\name{coreData-class}
\alias{coreData-class}
\title{A class to represent \code{data.frame}-like objects within cores}
\description{
This is a \code{VIRTUAL} class implementing spreadsheet-like behaviour that
emulates \emph{some} of the \code{data.frame} functionality.
}
\details{
Objects belonging to subclases of \code{coreData} must have equal length vectors
within slots, representing variables along a series of observations. This allows
to define formal subclasses that inherit this basic structure, being specific
about the type of data they should hold. Methods allow to subset these objects
by rows and columns, and to coerce them to \code{data.frame} or \code{list} objects if
needed. \code{[} method allows a single argument and subsets by rows, returning an
object of the same class as the one that is being subsetted. \code{[[} method allows
subsetting by rows and columns, returning a named list. Objects can be created
from a single \code{data.frame} or \code{list} if names match the corresponding slot
names.
}
\seealso{
Other core data classes: \code{\link{rawCore-class}},
  \code{\link{subCore-class}}
}
\concept{core data classes}
