% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{hdr}
\alias{hdr}
\alias{hdr.default}
\title{Compute highest density region}
\usage{
hdr(x, ...)

\method{hdr}{default}(value, prob, int = NULL)
}
\arguments{
\item{value}{A numeric vector of values for which the probability function
gives probabilities.}

\item{prob}{A numeric vector of probabilities of the same length as \code{value}}

\item{int}{A numeric vector describing the \emph{size} of the desired range, as a
percentage. The default, \code{c(95, 99)}, will give the intervals that account
for 95% and 99% of the probability.}
}
\description{
Computes the interval holding the desired probability density
for a probability distribution.
}
\details{
This is a generic function. The default method takes a \code{value} and a
\code{prob} vector and finds the range(es) within \code{value} holding the probability
density defined by \code{int}.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: 
}}

