% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{init}
\alias{init}
\title{Initialize objects within constructor functions}
\usage{
init(x, ...)
}
\arguments{
\item{x}{An object of the class to be initialized}
}
\value{
An object of the class being initialized, as modified by the
\code{init.className} method.
}
\description{
\code{init()} is a generic intended to \emph{initialize} objects within a
constructor function. Usage is only meaningful as a way to isolate special
operations needed for a certain class from the structure of the object
itself. This keeps the code more readable.
}
