% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{normalize}
\alias{normalize}
\title{Normalize a vector of probabilities}
\usage{
normalize(x, ...)
}
\arguments{
\item{x}{A numeric vector of probabilies, or an object from a class for which
a method has been defined.}
}
\value{
A numeric vector, or an object from a class for which a method has
been created.
}
\description{
Given a vector of probabilities, normalize ensures they sum up
to 1.
}
\details{
This is a generic function. Methods can be written for classes
holding probability distributions. The default method takes a vector \code{x} and
returns \code{x / sum(x)}. Specific methods should \code{normalize} the element holding
the probabilities within the object, and return the modified version of that
object.
}
