% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-pmf.r
\name{pmf}
\alias{pmf}
\title{Informal class to hold a probability mass function}
\usage{
pmf(x = NULL, value, prob)
}
\arguments{
\item{x}{a numeric vector. This optional argument is used to compute a \code{pmf}
from a sample, in which case this should be the only argument.}

\item{value}{a numeric vector containing the possible values within the \code{pmf}
support range.}

\item{prob}{a numeric vector containing the probability mass associated with
each \code{value}.}
}
\value{
a \code{list} with \code{$value} and \code{$prob} of class \code{pmf}.
}
\description{
A constructor function for objects of class \code{pmf}, either from
the values and associated probability mass or from a sample of the desired
distribution.
}
