% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rawCore.r
\docType{class}
\name{rawCore-class}
\alias{rawCore-class}
\title{a class to represent segmented cores from a series of boreholes}
\description{
Class \code{rawCore} holds a core in the form that it is usually
recovered when using a 'piston corer' or similiar tools that retrieve the
core piecewice. It represents the core as a series of \code{subCore}s, with an
internal depth scale, and provides methods to build a \code{core} object with a
composite depth scale for further analysis.
}
\section{Slots}{

\describe{
\item{\code{core}}{A length one character string with the name of the core. Should
match the \code{core} slot at any embeded \code{subCore} object.}

\item{\code{boreholes}}{A character vector with the names of the boreholes, or
something that can be coerced to.}

\item{\code{subCores}}{A list of objects of class \code{subCore}.}
}}

\seealso{
Other core data classes: \code{\link{coreData-class}},
  \code{\link{subCore-class}}
}
\concept{core data classes}
