% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schron-package.r
\docType{package}
\name{schron}
\alias{schron}
\alias{schron-package}
\title{Formal classes and methods for building chronologies simultaneously
for several sediment cores, and for dealing with sediment cores in general.}
\description{
\code{schron} uses available methods from \link[rbacon:rbacon]{rbacon::rbacon} and \link[Bchron:Bchron]{Bchron::Bchron} to build
chronologies for several sedimentary cores, using available information of
synchronic \emph{event layers} between cores to further constraint the
overall chronology. The package provides a set of formal classes and methods
tailored at computing a \emph{shared chronology}, but also suitable for general
manipulation of sediment core data or alike.
}
\details{
schron.
}
