% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_subCore.r
\docType{class}
\name{subCore-class}
\alias{subCore-class}
\title{a class to represent a core segment}
\description{
Class \code{subCore} represents a single segment from a core
recovered by using a 'piston corer' or any similar tool that retrieves a core
in a piecewise fashion.
}
\section{Slots}{

\describe{
\item{\code{core,borehole}}{A length one character vector with the name of the core
and the borehole. These should match the \code{core} slot at the enclosing
\code{\link{rawCore}} object, and one of the \code{boreholes}.}

\item{\code{subCore}}{A length one integer representing the sub core number. This
should be unique for each \code{subCore} within a single borehole.}

\item{\code{topDepth}}{A length one integer or double representing the depth from the
top of the sub core to the boring surface or datum.}

\item{\code{dLength,rLength}}{A length one integer or double providing the penetrated
(\emph{drilled}) and recovered depths, respectively. These values will be used to
correct for compression ---or decompression--- derived from the coring
exercise.}

\item{\code{An}}{object of class \code{\link{proxyData}} with arbitrary data on a unique depth
scale, arranged as a \emph{spresdsheet}.}
}}

\seealso{
Other core data classes: \code{\link{coreData-class}},
  \code{\link{rawCore-class}}
}
\concept{core data classes}
