% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{valid}
\alias{valid}
\title{Validate objects within constructor functions}
\usage{
valid(x, ...)
}
\arguments{
\item{x}{An object of the class to be validated.}
}
\value{
\code{TRUE} if successful.
}
\description{
\code{valid()} is a generic intended to validate objects from a class
within the constructor function for that class. Typical usage should be of
the form:\preformatted{constructor <- function (x, ...) {
  out <- structure(x, class = 'className')
  valid(out)
  out
}
}
}
\details{
\code{valid} is a generic function with no default. For it to work you
need to write a method for the class, usually as part of the class
definition. The \code{valid.className} method should return \code{TRUE} if all
conditions are met, and halt with informative error messages otherwise:\preformatted{valid.className <- fuction (x) {
  conditions <- with(x, c(
    given == expected,
     .
     .
     .
    )
  )
  if(all(conditions)) return(TRUE)
  stop('Informative error message')
}
}
}
